On branch timeFormat  Changes to be committed: 	modified:   TsIdxData.py 	modified:   ftArchPostProc.py
* Added the source data time format string as an argument.  This allows
flexible interpretation of the source data. There is a default setting
that interprets hours in the 12 hour format with an AM/PM designation at
the end of the string.  Without this functionality, date times starting
at 12 noon were being interpreted as 12 midnight as internally no format
string was used, and the date was being inferred and misinterpreted.