Changes to be committed:     modified:   TsIdxData.py     modified:   ftArchPostProc.py
Added in merge capability for -n Time Normalized Data, and should work
for 1-4 merge files and for data having more tags, the same tags with
additional times, or more tags and more times.

Needed:  More testing of merge files is needed, especially files 2-3 for
all data types.

TODO: Need to test/code for the -a archive data case when the same tag
is added but for different times.  I think currently the code just
merges columns, so the instrument would be repeated, rather than two
value columns for the same instrument being merged.

These are added for reference and to help test.
    new file:   phArchSample.csv
    new file:   vpiSampleSm1.csv
    new file:   vpiSampleSm2MoreTags.csv
    new file:   vpiSampleSm2MoreTagsTime.csv
    new file:   vpiSampleSm2MoreTime.csv