Changes to be committed: 	modified:   TsIdxData.py 	modified:   ftArchPostProc.py
*  Implemented the append function to the TsIdxData.py object. This
takes data and appends it to the member dataframe. Column names are
ignored, and a number of rows at the beginning of the data to be merged
can be ignored. This defaults to 1, anticipating a header row.  The data
to be merged assumes the column names used for the existing data. A
value query, if the member query is not none, and time filtering, if the
member time filters are not none, is performed before the data is merged.

*   Now ftArchPostProc keeps a list of instrument names that have been
created. If a file to be merged contains data for an instrument that
already has an TsIdxData object, or if an input file contains two data
sets for the same instrument, then the exiting data is appended to using the
TsIdxData.append() function. If a TsIdxData object does not exist in the
list of instrument names, the one is created. This bit works the same as
before this append functionality.