* Added a verbose option -v or --verbose which prints data that may be helpful for troubleshooting.
* After the list of instruments is built, the list is sorted by
the lower case version of the instrument name.  This is done late so
that sorting columns earlier in the process on large data sets does not
need to be done.

* Due to the sorting of the list, the merge/concat functions that build
the instrument data do not sort the columns, which at that point the
column names generaly refer to instrument names.

* Tested the merge functions for -t, -a, and -n data. All looks good.

* Updated the README.md

* Added archOut.csv as an example of an output file

* Uncommended out the bits that make the output file, and verified the
output file, in the correct format, is still created.

 Changes to be committed:
	modified:   README.md
	modified:   TrendSampleSm1RT.csv
	modified:   TsIdxData.py
	new file:   archOut.csv
	modified:   archiveSampleSm4.csv
	modified:   ftArchPostProc.py