On branch colTitles  Changes to be committed: 	modified:   ftArchPostProc.py
Drop Bias column in -n file types -- this is because it isn't a
process value, and this makes file merging when using a previous output
-n file as an input to a new -n file processing

Don't append 'value_' to instrument column names if they already start
with 'value'. Again this is for cleaner looking files when merging using
a -n out to -n in processing.